implemented first example for modular controller
l1 and l2, l3 still has to be done
